Adds friends components to separate folder
