+++ src/components/friends/FriendRequests.jsx
@@ -1,7 +1,7 @@
 import { useCollection } from 'react-firebase-hooks/firestore';
 import { collection, query, where, orderBy, doc, updateDoc, addDoc, serverTimestamp } from 'firebase/firestore';
 import { useAuthState } from 'react-firebase-hooks/auth';
-import { auth, db } from '../firebase';
+import { auth, db } from '../../firebase';
 
 const FriendRequests = ({ userId }) => {
   const [user] = useAuthState(auth);
